+++ src/localllm/__init__.py
@@ -9,10 +9,8 @@ from .dspy import LocalLLM, LocalChatAdapter, dspy_signature_defaults
 from .models import localllm_download_model
 from .utils import txt_locate_all, txt_locate, merge_spans
 from .config import localllm_connect
-from .converters import tif
 
 __all__ = [
     "localllm_connect",
     "LocalLLM", "localllm_download_model", "LocalChatAdapter", "dspy_signature_defaults",
-    "txt_locate_all", "txt_locate", "merge_spans",
-    "tif"]
+    "txt_locate_all", "txt_locate", "merge_spans"]